fix references to renamed src/core/utils, use paths relative to srcdir
