+++ src/com/opc_ua/opcua_client_config_parser.h
@@ -14,7 +14,7 @@
 #ifndef SRC_MODULES_OPC_UA_OPCUA_CLIENT_CONFIG_PARSER_H_
 #define SRC_MODULES_OPC_UA_OPCUA_CLIENT_CONFIG_PARSER_H_
 
-#include "utils/configFileParser.h"
+#include "core/util/configFileParser.h"
 #include "core/datatypes/forte_string.h"
 #include "open62541.h"
 #include <string>
