+++ src/com/opc_ua/opcua_helper.cpp
@@ -17,8 +17,8 @@
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
 #include "opcua_helper_gen.cpp"
 #endif
-#include "../../core/utils/parameterParser.h"
-#include "../../arch/devlog.h"
+#include "core/util/parameterParser.h"
+#include "arch/devlog.h"
 #include "convert_functions.h"
 #include "forte_date.h"
 #include "forte_struct.h"
