+++ src/core/cominfra/basecommfb.cpp
@@ -25,7 +25,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <devlog.h>
-#include "core/utils/string_utils.h"
+#include "core/util/string_utils.h"
 
 using namespace forte::com_infra;
