+++ src/core/ecet.cpp
@@ -18,8 +18,8 @@
 #include <fortenew.h>
 #include "ecet.h"
 #include "esfb.h"
-#include "utils/criticalregion.h"
-#include "../arch/devlog.h"
+#include "core/util/criticalregion.h"
+#include "arch/devlog.h"
 
 CEventChainExecutionThread::CEventChainExecutionThread() :
     CThread(){
